+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [4.1.1] - 2025-04-27
+
 ### Added
 
 - Added `federation.fep_ef61_gateway_enabled` configuration option.